backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.12.2
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1`
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
docker==7.1.0
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.58.4
future==1.0.0
gnureadline==8.2.13
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib==3.10.3
MAVProxy==1.8.71
numpy==2.3.0
openai==1.86.0
opencv-python==4.11.0.86
packaging==25.0
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
pymavlink==2.4.47
pynmeagps==1.0.50
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wxPython==4.2.3